// src/api/jikan/models/animeFull.ts
export interface AnimeFull {
  aired?: Daterange;
  /** Airing boolean */
  airing?: boolean;
  /** Whether the entry is pending approval on MAL or not */
  approved?: boolean;
  /**
   * Background
   * @nullable
   */
  background?: string | null;
  broadcast?: Broadcast;
  demographics?: MalUrl[];
  /**
   * Parsed raw duration
   * @nullable
   */
  duration?: string | null;
  /**
   * Episode count
   * @nullable
   */
  episodes?: number | null;
  explicit_genres?: MalUrl[];
  external?: AnimeFullExternalItem[];
  /**
   * Number of users who have favorited this entry
   * @nullable
   */
  favorites?: number | null;
  genres?: MalUrl[];
  images?: AnimeImages;
  licensors?: MalUrl[];
  /** MyAnimeList ID */
  mal_id?: number;
  /**
   * Number of users who have added this entry to their list
   * @nullable
   */
  members?: number | null;
  /**
   * Popularity
   * @nullable
   */
  popularity?: number | null;
  producers?: MalUrl[];
  /**
   * Ranking
   * @nullable
   */
  rank?: number | null;
  /**
   * Anime audience rating
   * @nullable
   */
  rating?: AnimeFullRating;
  relations?: AnimeFullRelationsItem[];
  /**
   * Score
   * @nullable
   */
  score?: number | null;
  /**
   * Number of users
   * @nullable
   */
  scored_by?: number | null;
  /**
   * Season
   * @nullable
   */
  season?: AnimeFullSeason;
  /**
   * Original Material/Source adapted from
   * @nullable
   */
  source?: string | null;
  /**
   * Airing status
   * @nullable
   */
  status?: AnimeFullStatus;
  streaming?: AnimeFullStreamingItem[];
  studios?: MalUrl[];
  /**
   * Synopsis
   * @nullable
   */
  synopsis?: string | null;
  theme?: AnimeFullTheme;
  themes?: MalUrl[];
  /**
   * Title
   * @deprecated
   */
  title?: string;
  /**
   * English Title
   * @deprecated
   * @nullable
   */
  title_english?: string | null;
  /**
   * Japanese Title
   * @deprecated
   * @nullable
   */
  title_japanese?: string | null;
  /**
   * Other Titles
   * @deprecated
   */
  title_synonyms?: string[];
  /** All titles */
  titles?: Title[];
  trailer?: TrailerBase;
  /**
   * Anime Type
   * @nullable
   */
  type?: AnimeFullType;
  /** MyAnimeList URL */
  url?: string;
  /**
   * Year
   * @nullable
   */
  year?: number | null;
}