// src/api/jikan/models/mangaFull.ts
export interface MangaFull {
  /** Whether the entry is pending approval on MAL or not */
  approved?: boolean;
  authors?: MalUrl[];
  /**
   * Background
   * @nullable
   */
  background?: string | null;
  /**
   * Chapter count
   * @nullable
   */
  chapters?: number | null;
  demographics?: MalUrl[];
  explicit_genres?: MalUrl[];
  external?: MangaFullExternalItem[];
  /**
   * Number of users who have favorited this entry
   * @nullable
   */
  favorites?: number | null;
  genres?: MalUrl[];
  images?: MangaImages;
  /** MyAnimeList ID */
  mal_id?: number;
  /**
   * Number of users who have added this entry to their list
   * @nullable
   */
  members?: number | null;
  /**
   * Popularity
   * @nullable
   */
  popularity?: number | null;
  published?: Daterange;
  /** Publishing boolean */
  publishing?: boolean;
  /**
   * Ranking
   * @nullable
   */
  rank?: number | null;
  relations?: MangaFullRelationsItem[];
  /**
   * Score
   * @nullable
   */
  score?: number | null;
  /**
   * Number of users
   * @nullable
   */
  scored_by?: number | null;
  serializations?: MalUrl[];
  /** Publishing status */
  status?: MangaFullStatus;
  /**
   * Synopsis
   * @nullable
   */
  synopsis?: string | null;
  themes?: MalUrl[];
  /**
   * Title
   * @deprecated
   */
  title?: string;
  /**
   * English Title
   * @deprecated
   * @nullable
   */
  title_english?: string | null;
  /**
   * Japanese Title
   * @deprecated
   * @nullable
   */
  title_japanese?: string | null;
  /**
   * Other Titles
   * @deprecated
   */
  title_synonyms?: string[];
  /** All Titles */
  titles?: Title[];
  /**
   * Manga Type
   * @nullable
   */
  type?: MangaFullType;
  /** MyAnimeList URL */
  url?: string;
  /**
   * Volume count
   * @nullable
   */
  volumes?: number | null;
}