// src/api/mangaupdates/models/seriesModelV1.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * MangaUpdates API
 * This API powers our website.
Most API functions are public and do not require an account. For user-based functions, you must register an account.
Currently, user registration must be done through our main website and is disabled via this API.

Download OpenAPI Specification: [openapi.yaml](openapi.yaml)

Please contact us at the following emails if you have questions:

* lambchopsil@mangaupdates.com
* manick@mangaupdates.com

## Warranties

MangaUpdates makes no warranties about service availability, correctness of the data, or anything else. The service is provided as is, and may change at any time.

## Acceptable Use Policy

* You will credit MangaUpdates when using data provided by this API.
* You will use reasonable spacing between requests so as not to overwhelm the MangaUpdates servers, and employ caching mechanisms when accessing data.
* You will NOT use MangaUpdates data or API in a way that will:
    * Deceive or defraud users
    * Assist or perform an illegal action
    * Create spam
    * Damage the database

We reserve the right to change this policy at any time.
 * OpenAPI spec version: 1.0.0
 */
import type { SeriesModelV1Admin } from './seriesModelV1Admin';
import type { SeriesModelV1Anime } from './seriesModelV1Anime';
import type { SeriesModelV1AssociatedItem } from './seriesModelV1AssociatedItem';
import type { SeriesModelV1AuthorsItem } from './seriesModelV1AuthorsItem';
import type { CategoriesModelV1 } from './categoriesModelV1';
import type { SeriesRecommendationsModelV1 } from './seriesRecommendationsModelV1';
import type { SeriesModelV1GenresItem } from './seriesModelV1GenresItem';
import type { ImageModelV1 } from './imageModelV1';
import type { TimeV1 } from './timeV1';
import type { SeriesModelV1PublicationsItem } from './seriesModelV1PublicationsItem';
import type { SeriesModelV1PublishersItem } from './seriesModelV1PublishersItem';
import type { SeriesModelV1Rank } from './seriesModelV1Rank';
import type { SeriesModelV1RelatedSeriesItem } from './seriesModelV1RelatedSeriesItem';
import type { SeriesModelV1Type } from './seriesModelV1Type';

export interface SeriesModelV1 {
  admin?: SeriesModelV1Admin;
  anime?: SeriesModelV1Anime;
  associated?: SeriesModelV1AssociatedItem[];
  authors?: SeriesModelV1AuthorsItem[];
  bayesian_rating?: number;
  categories?: CategoriesModelV1[];
  category_recommendations?: SeriesRecommendationsModelV1[];
  completed?: boolean;
  description?: string;
  forum_id?: number;
  genres?: SeriesModelV1GenresItem[];
  image?: ImageModelV1;
  last_updated?: TimeV1;
  latest_chapter?: number;
  licensed?: boolean;
  publications?: SeriesModelV1PublicationsItem[];
  publishers?: SeriesModelV1PublishersItem[];
  rank?: SeriesModelV1Rank;
  rating_votes?: number;
  recommendations?: SeriesRecommendationsModelV1[];
  related_series?: SeriesModelV1RelatedSeriesItem[];
  series_id?: number;
  status?: string;
  title?: string;
  type?: SeriesModelV1Type;
  url?: string;
  year?: string;
}